fix(game): service registration

+++ src/Core/HostingOptions.cs
@@ -4,7 +4,6 @@ namespace QuantumCore;
 
 public class HostingOptions
 {
-    [Required]
-    public int Port { get; set; }
+    [Required] public ushort Port { get; set; }
     public string? IpAddress { get; set; }
 }
